New files and maybe fixes
