+++ Task_1_1_1/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Wed Oct 05 20:12:00 NOVT 2022
+gradle.version=7.4
